Basic braindump of the watchers driven by a WAL table
The following is a braindump of watchers driven by a WAL table within
the database. Each entry is small in size and just identifies a
modification has occurred. Either a create, update or delete. Polling
the database every 100ms to see any changes should then be able to drive
our watchers based on any new WAL entries.

More work is required around subscribing and filtering the changes from
the WAL, so it's more ergonomic.